Fix search and cleanup

+++ src/replace-x.js
@@ -6,8 +6,12 @@ import xRegExp from 'xregexp';
 import sharedOptionsFactory from '../bin/shared-options-x';
 import 'colors';
 
-const sharedOptions = sharedOptionsFactory();
+const isNil = function isNil(value) {
+  /* eslint-disable-next-line lodash/prefer-is-nil */
+  return typeof value === 'undefined' || value === null;
+};
 
+const sharedOptions = sharedOptionsFactory();
 const hasOwnProp = Function.call.bind(Object.prototype.hasOwnProperty);
 
 const getFlags = function getFlags(options) {
@@ -100,7 +104,7 @@ const makeReplaceText = function makeReplaceText(options, canReplace) {
         return true;
       }
 
-      let replacement = String(options.replacement) || '$&';
+      let replacement = isNil(options.replacement) ? '$&' : String(options.replacement);
 
       if (!options.noColor) {
         replacement = replacement[options.color];
@@ -132,7 +136,7 @@ const makeReplaceText = function makeReplaceText(options, canReplace) {
   };
 };
 
-const makeReplacefile = function makeReplacefile(options, canReplace, replacizeText) {
+const makeReplacefile = function makeReplacefile(options, canReplace, replaceTextFn) {
   return function replaceFile(file) {
     fs.lstat(file, function lstat(error, stats) {
       if (error) {
@@ -160,7 +164,7 @@ const makeReplacefile = function makeReplacefile(options, canReplace, replacizeT
             throw err;
           }
 
-          const txt = replacizeText(text, file);
+          const txt = replaceTextFn(text, file);
 
           if (canReplace && txt !== null) {
             fs.writeFile(file, txt, function writeFile(e) {
@@ -185,7 +189,7 @@ const makeReplacefile = function makeReplacefile(options, canReplace, replacizeT
   };
 };
 
-const makeReplaceFileSync = function makeReplaceFileSync(options, canReplace, replaceText) {
+const makeReplaceFileSync = function makeReplaceFileSync(options, canReplace, replaceTextFn) {
   return function replaceFileSync(file) {
     const stats = fs.lstatSync(file);
 
@@ -201,12 +205,10 @@ const makeReplaceFileSync = function makeReplaceFileSync(options, canReplace, re
     }
 
     if (isFile) {
-      if (canReplace) {
-        const text = replaceText(fs.readFileSync(file, options.encoding), file);
+      const text = replaceTextFn(fs.readFileSync(file, options.encoding), file);
 
-        if (text !== null) {
-          fs.writeFileSync(file, text);
-        }
+      if (canReplace && text !== null) {
+        fs.writeFileSync(file, text);
       }
     } else if (stats.isDirectory() && options.recursive) {
       fs.readdirSync(file).forEach(function readdirSync(f) {
@@ -235,9 +237,9 @@ const replace = function replace(options) {
   opts.color = sharedOptions.color.choices.includes(opts.color) ? opts.color : 'cyan';
   opts.fileColor = sharedOptions.fileColor.choices.includes(opts.fileColor) ? opts.fileColor : 'yellow';
   const canReplace = !opts.preview && typeof opts.replacement !== 'undefined';
-  const replaceText = makeReplaceText(opts, canReplace);
+  const replaceTextFn = makeReplaceText(opts, canReplace);
   const makeReplaceFn = opts.async ? makeReplacefile : makeReplaceFileSync;
-  const replaceFileFn = makeReplaceFn(opts, canReplace, replaceText);
+  const replaceFileFn = makeReplaceFn(opts, canReplace, replaceTextFn);
 
   opts.paths.forEach(replaceFileFn);
 };
